First Commit
This is the first commit of my toy Zend PHP app.  It gets transaction
info and joins it with customer info and presents it to the user.
There is the ability to filter the results using jQuery datepickers as
well.  I did this for a technical exam for a job, but using Zend wasn't
required.  I just wanted to see what using Zend is like.